Moved HadoopConfiguration into ModifiableHadoopConfiguration and renamed
