+++ hadoop/config/HadoopConfiguration.java
@@ -19,8 +19,12 @@
 package grakn.core.hadoop.config;
 
 import com.google.common.base.Preconditions;
+import grakn.core.graph.diskstorage.configuration.ConfigElement;
+import grakn.core.graph.diskstorage.configuration.ConfigNamespace;
+import grakn.core.graph.diskstorage.configuration.ModifiableConfiguration;
 import grakn.core.graph.diskstorage.configuration.WriteConfiguration;
 import grakn.core.graph.diskstorage.util.time.Durations;
+import grakn.core.graph.graphdb.configuration.GraphDatabaseConfiguration;
 import org.apache.hadoop.conf.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,168 +33,193 @@ import java.lang.reflect.Constructor;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.time.temporal.TemporalUnit;
-import java.util.Map.Entry;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-public class HadoopConfiguration implements WriteConfiguration {
+public class HadoopConfiguration extends ModifiableConfiguration {
 
-    private static final Logger LOG = LoggerFactory.getLogger(HadoopConfiguration.class);
+    private final Configuration conf;
 
-    private final Configuration config;
-    private final String prefix;
-
-    public HadoopConfiguration(Configuration config) {
-        this(config, null);
+    private HadoopConfiguration(ConfigNamespace root, Configuration c) {
+        super(root, new HadoopWriteConfiguration(c));
+        this.conf = c;
     }
 
-    public HadoopConfiguration(Configuration config, String prefix) {
-        this.config = config;
-        this.prefix = prefix;
+    public static HadoopConfiguration of(ConfigNamespace root, Configuration c) {
+        Preconditions.checkNotNull(c);
+        return new HadoopConfiguration(root, c);
     }
 
-    @Override
-    public <O> O get(String key, Class<O> dataType) {
+    private static ModifiableConfiguration prefixView(HadoopConfiguration mc) {
+        HadoopWriteConfiguration prefixConf = new HadoopWriteConfiguration(mc.conf,
+                                                                           ConfigElement.getPath(JanusGraphHadoopConfiguration.GRAPH_CONFIG_KEYS, true) + ".");
+        return new ModifiableConfiguration(GraphDatabaseConfiguration.ROOT_NS, prefixConf);
+    }
 
-        String internalKey = getInternalKey(key);
+    public ModifiableConfiguration getJanusGraphConf() {
+        return prefixView(this);
+    }
 
-        if (null == config.get(internalKey)) {
-            return null;
+    public static class HadoopWriteConfiguration implements WriteConfiguration {
+
+        private static final Logger LOG = LoggerFactory.getLogger(HadoopWriteConfiguration.class);
+
+        private final Configuration config;
+        private final String prefix;
+
+        HadoopWriteConfiguration(Configuration config) {
+            this(config, null);
         }
 
-        if (dataType.isArray()) {
-            Preconditions.checkArgument(dataType.getComponentType() == String.class, "Only string arrays are supported: %s", dataType);
-            return (O) config.getStrings(internalKey);
-        } else if (Number.class.isAssignableFrom(dataType)) {
-            String s = config.get(internalKey);
-            return constructFromStringArgument(dataType, s);
-        } else if (dataType == String.class) {
-            return (O) config.get(internalKey);
-        } else if (dataType == Boolean.class) {
-            return (O) Boolean.valueOf(config.get(internalKey));
-        } else if (dataType.isEnum()) {
-            O[] constants = dataType.getEnumConstants();
-            Preconditions.checkState(null != constants && 0 < constants.length, "Zero-length or undefined enum");
-            String estr = config.get(internalKey);
-            for (O c : constants) {
-                if (c.toString().equals(estr)) {
-                    return c;
-                }
-            }
-            throw new IllegalArgumentException("No match for string \"" + estr + "\" in enum " + dataType);
-        } else if (dataType == Object.class) {
-            // Return String when an Object is requested
-            // Object.class must be supported for the sake of AbstractConfiguration's getSubset impl
-            return (O) config.get(internalKey);
-        } else if (Duration.class.isAssignableFrom(dataType)) {
-            // This is a conceptual leak; the config layer should ideally only handle standard library types
-            String s = config.get(internalKey);
-            String[] comps = s.split("\\s");
-            final TemporalUnit unit;
-            switch (comps.length) {
-                case 1:
-                    //By default, times are in milli seconds
-                    unit = ChronoUnit.MILLIS;
-                    break;
-                case 2:
-                    unit = Durations.parse(comps[1]);
-                    break;
-                default:
-                    throw new IllegalArgumentException("Cannot parse time duration from: " + s);
-            }
-            return (O) Duration.of(Long.valueOf(comps[0]), unit);
-        } else throw new IllegalArgumentException("Unsupported data type: " + dataType);
-    }
+        HadoopWriteConfiguration(Configuration config, String prefix) {
+            this.config = config;
+            this.prefix = prefix;
+        }
 
-    @Override
-    public Iterable<String> getKeys(String userPrefix) {
-        /*
-         * Is there a way to iterate over just the keys of a Hadoop Configuration?
-         * Iterating over Map.Entry is needlessly wasteful since we don't need the values.
-         */
+        @Override
+        public <O> O get(String key, Class<O> dataType) {
 
-        return StreamSupport.stream(config.spliterator(), false)
-                .map(Entry::getKey)
-                .filter(internalKey -> {
-                    String k = internalKey;
-                    if (null != prefix) {
-                        if (k.startsWith(prefix)) {
-                            k = getUserKey(k);
-                        } else {
-                            return false; // does not have the prefix
-                        }
+            String internalKey = getInternalKey(key);
+
+            if (null == config.get(internalKey)) {
+                return null;
+            }
+
+            if (dataType.isArray()) {
+                Preconditions.checkArgument(dataType.getComponentType() == String.class, "Only string arrays are supported: %s", dataType);
+                return (O) config.getStrings(internalKey);
+            } else if (Number.class.isAssignableFrom(dataType)) {
+                String s = config.get(internalKey);
+                return constructFromStringArgument(dataType, s);
+            } else if (dataType == String.class) {
+                return (O) config.get(internalKey);
+            } else if (dataType == Boolean.class) {
+                return (O) Boolean.valueOf(config.get(internalKey));
+            } else if (dataType.isEnum()) {
+                O[] constants = dataType.getEnumConstants();
+                Preconditions.checkState(null != constants && 0 < constants.length, "Zero-length or undefined enum");
+                String estr = config.get(internalKey);
+                for (O c : constants) {
+                    if (c.toString().equals(estr)) {
+                        return c;
                     }
-                    return k.startsWith(userPrefix);
-                })
-                .map(internalKey -> {
-                    String userKey = getUserKey(internalKey);
-                    Preconditions.checkState(userKey.startsWith(userPrefix));
-                    return userKey;
-                })
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public void close() {
-        // nothing to do
-    }
-
-    @Override
-    public <O> void set(String key, O value) {
-
-        String internalKey = getInternalKey(key);
-        Class<?> dataType = value.getClass();
-
-        if (dataType.isArray()) {
-            Preconditions.checkArgument(dataType.getComponentType() == String.class, "Only string arrays are supported: %s", dataType);
-            config.setStrings(internalKey, (String[]) value);
-        } else if (Number.class.isAssignableFrom(dataType)) {
-            config.set(internalKey, value.toString());
-        } else if (dataType == String.class) {
-            config.set(internalKey, value.toString());
-        } else if (dataType == Boolean.class) {
-            config.setBoolean(internalKey, (Boolean) value);
-        } else if (dataType.isEnum()) {
-            config.set(internalKey, value.toString());
-        } else if (dataType == Object.class) {
-            config.set(internalKey, value.toString());
-        } else if (Duration.class.isAssignableFrom(dataType)) {
-            // This is a conceptual leak; the config layer should ideally only handle standard library types
-            String millis = String.valueOf(((Duration) value).toMillis());
-            config.set(internalKey, millis);
-        } else throw new IllegalArgumentException("Unsupported data type: " + dataType);
-    }
-
-    @Override
-    public void remove(String key) {
-        config.unset(getInternalKey(key));
-    }
-
-    private <O> O constructFromStringArgument(Class<O> dataType, String arg) {
-        try {
-            Constructor<O> ctor = dataType.getConstructor(String.class);
-            return ctor.newInstance(arg);
-            // ReflectiveOperationException is narrower and more appropriate than Exception, but only @since 1.7
-            //} catch (ReflectiveOperationException e) {
-        } catch (Exception e) {
-            LOG.error("Failed to parse configuration string \"{}\" into type {} due to the following reflection exception", arg, dataType, e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    private String getInternalKey(String userKey) {
-        return null == prefix ? userKey : prefix + userKey;
-    }
-
-    private String getUserKey(String internalKey) {
-        String k = internalKey;
-
-        if (null != prefix) {
-            Preconditions.checkState(k.startsWith(prefix), "key %s does not start with prefix %s", internalKey, prefix);
-            Preconditions.checkState(internalKey.length() > prefix.length());
-            k = internalKey.substring(prefix.length());
+                }
+                throw new IllegalArgumentException("No match for string \"" + estr + "\" in enum " + dataType);
+            } else if (dataType == Object.class) {
+                // Return String when an Object is requested
+                // Object.class must be supported for the sake of AbstractConfiguration's getSubset impl
+                return (O) config.get(internalKey);
+            } else if (Duration.class.isAssignableFrom(dataType)) {
+                // This is a conceptual leak; the config layer should ideally only handle standard library types
+                String s = config.get(internalKey);
+                String[] comps = s.split("\\s");
+                final TemporalUnit unit;
+                switch (comps.length) {
+                    case 1:
+                        //By default, times are in milli seconds
+                        unit = ChronoUnit.MILLIS;
+                        break;
+                    case 2:
+                        unit = Durations.parse(comps[1]);
+                        break;
+                    default:
+                        throw new IllegalArgumentException("Cannot parse time duration from: " + s);
+                }
+                return (O) Duration.of(Long.valueOf(comps[0]), unit);
+            } else throw new IllegalArgumentException("Unsupported data type: " + dataType);
         }
 
-        return k;
+        @Override
+        public Iterable<String> getKeys(String userPrefix) {
+            /*
+             * Is there a way to iterate over just the keys of a Hadoop Configuration?
+             * Iterating over Map.Entry is needlessly wasteful since we don't need the values.
+             */
+
+            return StreamSupport.stream(config.spliterator(), false)
+                    .map(Map.Entry::getKey)
+                    .filter(internalKey -> {
+                        String k = internalKey;
+                        if (null != prefix) {
+                            if (k.startsWith(prefix)) {
+                                k = getUserKey(k);
+                            } else {
+                                return false; // does not have the prefix
+                            }
+                        }
+                        return k.startsWith(userPrefix);
+                    })
+                    .map(internalKey -> {
+                        String userKey = getUserKey(internalKey);
+                        Preconditions.checkState(userKey.startsWith(userPrefix));
+                        return userKey;
+                    })
+                    .collect(Collectors.toList());
+        }
+
+        @Override
+        public void close() {
+            // nothing to do
+        }
+
+        @Override
+        public <O> void set(String key, O value) {
+
+            String internalKey = getInternalKey(key);
+            Class<?> dataType = value.getClass();
+
+            if (dataType.isArray()) {
+                Preconditions.checkArgument(dataType.getComponentType() == String.class, "Only string arrays are supported: %s", dataType);
+                config.setStrings(internalKey, (String[]) value);
+            } else if (Number.class.isAssignableFrom(dataType)) {
+                config.set(internalKey, value.toString());
+            } else if (dataType == String.class) {
+                config.set(internalKey, value.toString());
+            } else if (dataType == Boolean.class) {
+                config.setBoolean(internalKey, (Boolean) value);
+            } else if (dataType.isEnum()) {
+                config.set(internalKey, value.toString());
+            } else if (dataType == Object.class) {
+                config.set(internalKey, value.toString());
+            } else if (Duration.class.isAssignableFrom(dataType)) {
+                // This is a conceptual leak; the config layer should ideally only handle standard library types
+                String millis = String.valueOf(((Duration) value).toMillis());
+                config.set(internalKey, millis);
+            } else throw new IllegalArgumentException("Unsupported data type: " + dataType);
+        }
+
+        @Override
+        public void remove(String key) {
+            config.unset(getInternalKey(key));
+        }
+
+        private <O> O constructFromStringArgument(Class<O> dataType, String arg) {
+            try {
+                Constructor<O> ctor = dataType.getConstructor(String.class);
+                return ctor.newInstance(arg);
+                // ReflectiveOperationException is narrower and more appropriate than Exception, but only @since 1.7
+                //} catch (ReflectiveOperationException e) {
+            } catch (Exception e) {
+                LOG.error("Failed to parse configuration string \"{}\" into type {} due to the following reflection exception", arg, dataType, e);
+                throw new RuntimeException(e);
+            }
+        }
+
+        private String getInternalKey(String userKey) {
+            return null == prefix ? userKey : prefix + userKey;
+        }
+
+        private String getUserKey(String internalKey) {
+            String k = internalKey;
+
+            if (null != prefix) {
+                Preconditions.checkState(k.startsWith(prefix), "key %s does not start with prefix %s", internalKey, prefix);
+                Preconditions.checkState(internalKey.length() > prefix.length());
+                k = internalKey.substring(prefix.length());
+            }
+
+            return k;
+        }
     }
 }
